slack-bot/infrastructure/pyproject.toml
[project]
name = "infrastructure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.100.0",
    "boto3>=1.40.7",
    "constructs>=10.0.0",
    "python-dotenv>=1.0.0",
]
